+++ packages/sdk-metrics/test/aggregator/ExponentialHistogram.test.ts
@@ -714,6 +714,7 @@ describe('ExponentialHistogramAggregation', () => {
               scale: 1,
               zeroCount: 0,
             },
+            exemplars: undefined,
           },
         ],
       };
@@ -765,6 +766,7 @@ describe('ExponentialHistogramAggregation', () => {
               scale: 1,
               zeroCount: 0,
             },
+            exemplars: undefined,
           },
         ],
       };