+++ CHANGELOG.md
@@ -15,10 +15,11 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 ### :rocket: Features
 
 * feat(sdk-logs): implement log creation metrics  [#6433](https://github.com/open-telemetry/opentelemetry-js/pull/6433) @anuraaga
+* feat(sdk-metrics): wire exemplar support into metrics pipeline [#6483](https://github.com/open-telemetry/opentelemetry-js/pull/6483) @CharlieTLe
 
 ### :bug: Bug Fixes
 
-fix(opentelemetry-resources): do not discard OTEL_RESOURCE_ATTRIBUTES when it contains empty kv pairs
+* fix(opentelemetry-resources): do not discard OTEL_RESOURCE_ATTRIBUTES when it contains empty kv pairs
 
 ### :books: Documentation