+++ experimental/CHANGELOG.md
@@ -40,6 +40,7 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 * feat(configuration): export interfaces required in other packages [#6462](https://github.com/open-telemetry/opentelemetry-js/pull/6462) @maryliag
 * feat(configuration): set MeterProvider on sdk start [#6463](https://github.com/open-telemetry/opentelemetry-js/pull/6463) @maryliag
 * feat(configuration): export interfaces required in other packages [#6507](https://github.com/open-telemetry/opentelemetry-js/pull/6507) @maryliag
+* feat(otlp-transformer): serialize exemplars on all metric data point types [#6483](https://github.com/open-telemetry/opentelemetry-js/pull/6483) @CharlieTLe
 
 ### :bug: Bug Fixes